requirements.txt
lxml
scraperwiki_local
cssselect
html2text
requests
BeautifulSoup
